#104: made sitesParser code reusable across Home page and Resources page
